fixed incorrect validation error claiming filePaths missing in game arcs
The RE-Engine wrapper is configured to extend the hash map/file list as
it encounters new mod files which aren't available in the pre-generated
file list (included with the game extension).

Unfortunately there was a bug causing the BMS scripts to use the
pre-generated hash map/file list which means that QBMS would faile to
find and/or invalidate file paths which the game archives actually
contained!

fixes nexus-mods/vortex#9147