fixed pending deployment dialog launching game while validation op runs
Previously if the user had forgotten to deploy his mods and decided to
run the game, the pending deployment dialog would allow the user to
choose to deploy the mods before running the game - unfortunately the
invalidation operation would still be in progress when Vortex would
finish the deployment and run the game.

Also improved the error handler.